Guaranteed sequential message handling
Messages are now fully processed in the order they came in.

Only one service can connect to the publication of a replication slot. When
services are scaled, the first one will succeed to connect but the
others will fail. There is now a new setting `restartDelaySlotInUse` to
define the delay before trying to connect again if the replication slot
is in use.